Add java/dotnet output in individual subfolders
Historically the output of each library has tried to overwrite the other.
Over time this has been very difficult to manage, and gets worse as
more images are added and the libraries diverge.

We will store each library's output in a different file (as in this
commit) and use other means to track differences in output.