Document Java version used to produce output
I recently installed OpenJDK to my machine which caused the output for several files to differ significantly, from date formatting to error messages. We cannot assume a specific version is installed on the user's PC, so I document the version I use here so I can copy it into the `FileName` property value when I need to run it.